An updated main.c file
This is a commit performed in a forked holbertonschool/your_first_code repository as instructed in the GitHub challenge of Holberton school Software engineering program. The instructions of the challenge are simple and straigh forward. To change the display results of the code to "Hello School" and modify the number of repeats of the message to 32